pyproject.toml
[project]
name = "smart-complaint-classifier-public-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "python-dotenv>=0.21.0",
    "google-generativeai>=0.3.2",
    "pandas>=2.2.2",
    "streamlit>=1.36.0",
    "plotly>=5.15.0",
]
